-Added motion and shooting to ship

+++ AMBroSIA/src/game/Projectile.java
@@ -9,7 +9,7 @@ public class Projectile extends MapObject {
     /**
      * The velocity of Projectile. The default is 10.
      */
-    public final static int[] PROJECTILE_VELOCITY = {10, 10};
+    public final static float[] PROJECTILE_VELOCITY = {10, 10};
     /**
      * The length of time that Projectile lasts. The default is 5.
      */
@@ -36,7 +36,7 @@ public class Projectile extends MapObject {
      * @param coordinates   
      * @param gameState     
      */
-    public Projectile(Ship ship, float[] velocity, int heading, int[] coordinates, GameState gameState){     
+    public Projectile(Ship ship, float[] velocity, float heading, int[] coordinates, GameState gameState){     
         super(velocity, heading, coordinates, 0, gameState);
         this.ttl = TIME_TO_LIVE;
